refactor: rename SubmitBtns to SubmitButtons
