+++ app/components/Form/index.ts
@@ -3,4 +3,4 @@ export { default as ApplicationForm } from './ApplicationForm';
 export { default as CalculationForm } from './CalculationForm';
 export { default as FormBase } from './FormBase';
 export { default as Label } from './Label';
-export { default as SubmitBtns } from './SubmitBtns';
+export { default as SubmitButtons } from './SubmitButtons';
